Extract TIFFReader to separate source file, add unit tests
https://github.com/brawer/osmviews/issues/30
